+++ kwarray/arrayapi.py
@@ -25,7 +25,6 @@ Example:
     >>> data2 = data1.numpy()
 """
 from __future__ import absolute_import, division, print_function, unicode_literals
-import torch
 import numpy as np
 import ubelt as ub
 import six
@@ -33,7 +32,13 @@ from functools import partial
 from distutils.version import LooseVersion
 
 
-_TORCH_HAS_BOOL_COMP = LooseVersion(torch.__version__) >= LooseVersion('1.2.0')
+try:
+    import torch
+except Exception:
+    torch = None
+    _TORCH_HAS_BOOL_COMP = False
+else:
+    _TORCH_HAS_BOOL_COMP = LooseVersion(torch.__version__) >= LooseVersion('1.2.0')
 
 
 def _get_funcname(func):
@@ -87,7 +92,7 @@ class _ImplRegistry(object):
             numpy_func = getattr(NumpyImpls, key)
             torch_func = getattr(TorchImpls, key)
             def func(data, *args, **kwargs):
-                if torch.is_tensor(data):
+                if torch is not None and torch.is_tensor(data):
                     return torch_func(data, *args, **kwargs)
                 elif isinstance(data, np.ndarray):
                     return numpy_func(data, *args, **kwargs)
@@ -196,6 +201,7 @@ class TorchImpls(object):
         Concatenates along axis=0 if inputs are are 1-D otherwise axis=1
 
         Example:
+            >>> # xdoctest: +REQUIRES(module:torch)
             >>> datas1 = [torch.arange(10), torch.arange(10)]
             >>> datas2 = [d.numpy() for d in datas1]
             >>> ans1 = TorchImpls.hstack(datas1)
@@ -214,6 +220,7 @@ class TorchImpls(object):
         and then concats along axis=0.
 
         Example:
+            >>> # xdoctest: +REQUIRES(module:torch)
             >>> datas1 = [torch.arange(10), torch.arange(10)]
             >>> datas2 = [d.numpy() for d in datas1]
             >>> ans1 = TorchImpls.vstack(datas1)
@@ -253,6 +260,7 @@ class TorchImpls(object):
     def compress(data, flags, axis=None):
         """
         Example:
+            >>> # xdoctest: +REQUIRES(module:torch)
             >>> import kwarray
             >>> data = torch.rand(10, 4, 2)
             >>> impl = kwarray.ArrayAPI.coerce(data)
@@ -307,6 +315,7 @@ class TorchImpls(object):
 
         Example:
             >>> # xdoctest: +SKIP
+            >>> # xdoctest: +REQUIRES(module:torch)
             >>> data = torch.arange(10)[:, None]
             >>> ans1 = ArrayAPI.tile(data, [1, 2])
             >>> ans2 = ArrayAPI.tile(data.numpy(), [1, 2])
@@ -314,6 +323,7 @@ class TorchImpls(object):
 
         Doctest:
             >>> # xdoctest: +SKIP
+            >>> # xdoctest: +REQUIRES(module:torch)
             >>> shapes = [(3,), (3, 4,), (3, 5, 7), (1,), (3, 1, 3)]
             >>> for shape in shapes:
             >>>     data = torch.rand(*shape)
@@ -402,6 +412,7 @@ class TorchImpls(object):
     def transpose(data, axes):
         """
         Example:
+            >>> # xdoctest: +REQUIRES(module:torch)
             >>> data1 = torch.rand(2, 3, 5)
             >>> data2 = data1.numpy()
             >>> res1 = ArrayAPI.transpose(data1, (2, 0, 1))
@@ -466,6 +477,7 @@ class TorchImpls(object):
     def max(data, axis=None):
         """
         Example:
+            >>> # xdoctest: +REQUIRES(module:torch)
             >>> data1 = torch.rand(5, 5, 5, 5, 5, 5)
             >>> data2 = data1.numpy()
             >>> res1 = ArrayAPI.max(data1)
@@ -505,6 +517,7 @@ class TorchImpls(object):
     def maximum(data1, data2, out=None):
         """
         Example:
+            >>> # xdoctest: +REQUIRES(module:torch)
             >>> data1 = torch.rand(5, 5)
             >>> data2 = torch.rand(5, 5)
             >>> result1 = TorchImpls.maximum(data1, data2)
@@ -517,6 +530,7 @@ class TorchImpls(object):
     def minimum(data1, data2, out=None):
         """
         Example:
+            >>> # xdoctest: +REQUIRES(module:torch)
             >>> data1 = torch.rand(5, 5)
             >>> data2 = torch.rand(5, 5)
             >>> result1 = TorchImpls.minimum(data1, data2)
@@ -644,6 +658,7 @@ class TorchImpls(object):
     def round(data, decimals=0, out=None):
         """
         Example:
+            >>> # xdoctest: +REQUIRES(module:torch)
             >>> import kwarray
             >>> rng = kwarray.ensure_rng(0)
             >>> np_data = rng.rand(10) * 100
@@ -815,6 +830,8 @@ class NumpyImpls(object):
 
     @_numpymethod(func_type='data_func')
     def tensor(data, device=ub.NoParam):
+        if torch is None:
+            raise Exception('torch is not available')
         data = torch.from_numpy(np.ascontiguousarray(data))
         if device is not ub.NoParam:
             data = data.to(device)
@@ -841,7 +858,7 @@ class NumpyImpls(object):
         """
         Cast data into a numpy representation
         """
-        if isinstance(data, torch.Tensor):
+        if torch is not None and isinstance(data, torch.Tensor):
             data = data.cpu().numpy()
         return np.asarray(data, dtype=dtype)
 
@@ -902,6 +919,7 @@ class ArrayAPI(object):
 
     Example:
         >>> # Use the easy-to-use, but inefficient array api
+        >>> # xdoctest: +REQUIRES(module:torch)
         >>> take = ArrayAPI.take
         >>> np_data = np.arange(0, 143).reshape(11, 13)
         >>> pt_data = torch.LongTensor(np_data)
@@ -914,6 +932,7 @@ class ArrayAPI(object):
 
     Example:
         >>> # Use the easy-to-use, but inefficient array api
+        >>> # xdoctest: +REQUIRES(module:torch)
         >>> compress = ArrayAPI.compress
         >>> np_data = np.arange(0, 143).reshape(11, 13)
         >>> pt_data = torch.LongTensor(np_data)
@@ -926,6 +945,7 @@ class ArrayAPI(object):
 
     Example:
         >>> # Use ArrayAPI to coerce an identical API that doesnt do type checks
+        >>> # xdoctest: +REQUIRES(module:torch)
         >>> import kwarray
         >>> np_data = np.arange(0, 15).reshape(3, 5)
         >>> pt_data = torch.LongTensor(np_data)
@@ -955,7 +975,7 @@ class ArrayAPI(object):
             data (ndarray | Tensor): data to be operated on
 
         """
-        if torch.is_tensor(data):
+        if torch is not None and torch.is_tensor(data):
             return TorchImpls
         else:
             return NumpyImpls